Add docs on authorising URL ownership for Indexing API access
We've had the following error crop up several times
"Google::Apis::ClientError: forbidden: Permission denied. Failed to verify the URL ownership."

We believe that ownership access was lost as part of the move from
Google Analytics to Google Tag Manager for the
"teachingjobs@digital.education.gov.uk" account.  Delegated access
for the API is removed if the delegator loses access.

These docs serve to simplify the process of fixing the error
if it crops up again.

Rollbar details:
https://rollbar.com/dfe/teacher-vacancies/items/219/