Rename MAX_SALARY_ALLOWED to MAX_SALARY_LIMIT
As part of the error content changes being done across the site, one of the desired changes is to display 'Minimum salary/Maximum salary must be less than £200,000'. This means that the upper limit that a user would be allowed to enter would be £199,999.99.

The error message generated uses the MAX_SALARY_ALLOWED constant as part of the message. Since this would still need to be £200,000 for the desired change, the name of the constant has been changed and where it has been implemented has be altered to account for the new validation.