Refactor #clear_all_rows method in spreadsheet writer
The spreadsheet writer class uses #num_rows method from the google drive gems' to find the last empty row when clearing the spreadsheet. This method is expensive as it maps all the data from the spreadsheet to find the first empty row. This causes a failure in clearing and updating the sheet when there are large amount of data. Replacing #num_rows method to #max_rows solves this problem. https://www.rubydoc.info/gems/google_drive/2.1.2/GoogleDrive%2FWorksheet:num_rows