Fix job alerts for minimum salary
The job alerts use Elasticsearch queries to return any matches for a search filter to use in the daily job alerts.
The salary query used for the job alerts also excluded vacancies with salary ranges where the minimum salary's lower than the filter term but the maximum salary's higher than the filter term.
This query also has some old behaviour: it can use a value for Maximum Salary, if given. However, Maximum Salary is no longer a filter option on the site.

We want the results given to user via the job alerts to match the results they would now get when they search site using the same filter.

This commit modifies the salary query for job alerts to instead:
- Return vacancies that have a maximum salary that is higher than the filter value
- Return vacancies that do NOT have a specified maxium salary, regardless of their minimum salary value
- Not reference the Maximum Salary filter value.