Refactor Vacancy Alert Search Builder
The class VacancyAlertSearchBuilder already inherits from the VacancySearchBuilder class.

Since the query for filtering by minimum salary should behave the same for displaying matching vacancies in either the job alerts or on the site.
This commit removes duplicated code for the salary query in the Vacancy Alert Search Builder.
It also removes code references to the maximum salary filter.