(chore) add link to feedback nudge email
We wanted to make it more convenient for our hiring staff users to
provide feedback regarding their expired vacancies. A quick win to make
it more convenient for them was to include links to either sign in or
to be taken to the expired vacancies tab if they are already signed in.
This would give hiring staff a quicker route to providing feedback.