Re-raise errors in dsi_api_response_to_spreadsheet usecase
Handled exeptions don't show up on Rollbar. As suggested by Adrian Clay, leaving these exeptions unhandled would allow us to know when a scheduled job failed. This commit changes the behaviour of the usecase, when there is an error on a given page, the spreadsheet writer task will terminate on the page.