(fix) fix hidden field to use array
When redirecting with a sort order after choosing a phase the search
would error as the hidden field was converting the phases array to a
string so the new URL would contain phases=secondary rather than
phases[]=secondary. This would cause an error when the phases param was
JSON parsed. We fixed it thanks to this answer on Stack Overflow:

https://stackoverflow.com/questions/28139318/pass-an-array-of-ids-back-from-a-view-to-a-rails-controller