Disable the exporting of "search_event" and "interest_expression" data.
These exports have been failing for several months now, causing out
of memory errors in the container they are run on:

```
Nov 25 01:05:43 Production ecs-agent 2019-11-25T01:05:42Z
[INFO] process within container 18c8e417918716434c0fccd70e72074e9ee6d8daba95b49f93cf3cde500e624a
(name: "ecs-tvs2_production_worker-121-tvs2productionweb-92aaede1cfc2faf49001") died due to OOM
```

We have agreed to continue collecting this information as it may
be useful for future analysis, but in the short term there is no
need for fixing the exporter.  Instead we could look at fixing
this once we move over to an alternative to Google Sheets.