Disable/enable radius filter selection when searching by location
When a user has performed a location category search, we want to disable the radius filter as they would receive all vacancies that fall within a given category for their location.
If the user wants to redefine their search e.g use a postcode or different location, the radius filter would be re-enabled upon clicking the location text field in JS-enabled browsers.
For JS-Disabled browsers, the user would have to click on the search button, the new search results page would have the radius filter enabled.