(chore) remove turbolinks from project
We didn't really feel like turbolinks was adding a lot of value to our
project and were constantly having to add workarounds to get it to work
with Google services that we're using. We decided that we would remove
turbolinks and see if this had any adverse effects on the site.