Add '&.' on values that could return nil
`min_pay_scale` and `max_pay_scale` are attributes of vacancies. If they are present, we want to get the label e.g `max_pay_scale.label` but this errors if `max_pay_scale` is nil.
This commit adds the safe operator on `min_pay_scale` and `max_pay_scale` so that nil is returned rather than an error.