Refactor test to make it less specific re. content
I decided to restrict the spec so it only looks for the prepostion and the location. My thinking is that trying to match
the full copy would ultimately make the test more fragile and does not really reflect the most important point of this
ticket - the *location* of the job for SEO purposes.