(chore) remove get more info clicks column
This currently isn't sending hiring staff the message that's desired,
especially since there are other ways in which jobseekers can engage
with a job posting e.g. They can see the job listing and call the school
directly instead of clicking the get more info button. For now we're
going to remove it until we can give hiring staff a better way of
measuring engagement with their job listings.