Export the most recent Vacancy data
The `run!` method retrieves data from the database and inserts it into an auto-created table in big query.
We want to keep the data in big query in sync with the data in the database.

The approach taken was to delete and recreate the vacancy big query table each time the `run!` command is executed.