(fix) fix flaky authorisation test
This test flaked once but when I reran it with the same seed it didn't
fail again. I was unable to get the test to fail again no matter how
much I reran the test. The initial failure was caused by a time
difference so I just changed the test to use Timecop to remove the risk
of future flakiness.