Replace #add_header with #inspect_request
One of the problems of chaning over to `selenium-webdriver` is there is no corollary to `Poltergeist#add_header` and the
maintainers have said there will never be. In order to avoid having to drastcially refactor the test that uses this
call, I ported the lightweight middleware that GitLab used to solve the same problem. Details can be found here:

https://about.gitlab.com/blog/2017/12/19/moving-to-headless-chrome/

There was only one test using it, but I can see the utility moving forward-especially for the middelware and examples
that wait on pages to complete loading.