Re-enable radius when using current location link
Previously users on a location category page were unable to
use the "Radius" field after clicking the "use your current location"
link as the field was disabled on the location catogory page load.

We want users to be able to select a non-default radius if they click
the link to use their current location.