(chore) remove expiry behaviour from job alert subscriptions
We weren't telling our users that their subscriptions to job alerts
would expire after 6 months. We removed this as we learned that we could
retain email addresses for as long as a user wanted to remain subscribed
without needing any extra permission - we also think that users were
probably subscribing under the assumption that they would have to
unsubscribe if they wanted to stop receiving job alert emails, so this
change should align us to those expectations.