(chore) select multiple working patterns
We changed the working patterns from a select box to a collection of
check boxes as user research revealed that some of our users wanted to
be able to search for multiple working pattern jobs - especially if they
were coming back to work after an extended period. This was because they
wanted to be able to see more of their potential options without having
to do multiple searches to see all the options they were interested in.

We also edited the tests to reflect the new behaviour.