Filter out web crawlers visits from tracking.
There was a large discrepancy found between the total number of page views, in the database and GA, for vacancies published in Autumn 2019.
Investigating the case based on the hypothesis that the value in the database was incorrect led to the discovery that our application’s in-built tracker did not ignore visits by web crawlers.
Given the number of both good and bad web crawlers, this could have contributed to inflating the total number of page views in the database.

This commit adds another condition for the tracker to verify before attempting to log the visit as a page view.
Any suggestion or feedback would be appreciated.