(fix) add rake script to update subscriptions to new working pattern
Subscriptions that currently exist in the database have the old style
of working pattern based on only being able to select one, however the
subscriptions code was update to handle arrays for working patterns as
we wanted users to be be able to search for multiple working patterns.
The result is that subscriptions created before the mutliple working
patterns change would no longer receive emails.

This change adds a rake task to update all the subscriptions in the
database to make themcompatible with the working pattern changes
(specifically those relatedto job alerts).