(fix) index homepage when visiting with root path
When we switched over to the new homepage we introduced a bug where we
set a noindex header on the homepage - this was because the homepage was
accessible through two paths, the root path (/) and the home page path
(/pages/home), but we were setting the noindex header for both paths.
This was bad as it would negatively impact our visibility on search
engines.

This commit allows the home page to be indexed with the root path, but a
noindex header is still set on the home page path (/pages/home). Some
tests were also added to the test suite to test for this.