(chore) initial import spike
This code imports files from our google cloud storage bucket to our big
query but uses a hard coded list of files and logs things with puts. We
mainly wanted to spike this out to get a better understanding of how it
worked and future commits will improve it and wrap some tests around the
code.