(chore) get all table CSVs uploaded by exporter
This change means that we use the `ExportTablesToCloudStorage::TABLES`
defined in export_tables_to_cloud_storage.rb as the source of truth for
the CSV files that will exist in the cloud storage. We changed the
logging to log things to rollbar since this will allow us to see any
issues with the CSV importing in our Slack channel for logging.