Google drive api av spike (#1279)
* Add document for information about Google Drive

As I am starting by investigating AV on the Google Drive API, that is
how the document begins.

* Expand documentation with outcome of current work

* Add details about alternative malware test files

* Add serviceAccount.json to .gitignore

This is mainly for ease of testing various workflows for antivirus
alerting and probably won't need to be pulled in the main repo. I'm
checking it in just-in-case.

* Outline a quick way to check for viruses

This isn't meant to be exhaustive, or even necessairly the best way.
However, it is quick, simple, and uses gems we're already using.