Add 'key info' properties to the updating of school data.
We changed the update_school_data.rb to set the new information
that we want based on the migration in the earlier commit. This
information will be used by the business analysts for school-based
metrics. We also want to surface some of this new information to
job-seekers to help them make a decision.