Extract religious character logic from view to model
The front end logic was throwing an error on edge because
schools in the edge database did not have a gias_data
attribute. That will be the case for schools in
production too, before they are updated by the job
which attaches GIAS data to their records. So to
prevent errors being thrown, .has_religious_character?
checks if the method gias_data exists.