Alias gias_data to data
This allows the updated BigQuery uploader to process `School` exactly as
it is processing the other records with `json` attributes. If a school
is missing the necessary json the processer will use schema information
from the other records to ensure that the school in question has the
correct columns with nulled values.

At the time of commit, there are more than 1800 out of about
27,000 shcool that are missing gias data.