Alter the processor to use JSON instead of CSV
This is a large change and will necessitate changing or removing some of
the exiting tests.

BigQuery works best if it has a flat, clean data structure to ingest.
The two-tierd nature of the exiting `data` attributes, combined with
their tendency to contain incomplete records, makes the import process
fragile and prone to errors. By normalizing the structure of the json
data across all related records, hoisting it to the same level as the
attributes and prefixing with a `data_` identifier, it becomes simpler
to import to BigQuery and easier to maintain. BigQuery can also
intelligently cast the column types, which saves Performance Analytics
from having to work it out in their BigQuery SQL.

The commit is less broken down than I would have liked, but I took this
on after a late delivery and needed to get the changes live ASAP to
unblock Performance Analytics. It sensible TODO would be to refactor it
when we have time to look at the overall data structuring.