Exclude the BigQuery lib from AbcSize metric check
Data cleaning on multiple sources is always an involved process and the
tradeoff is between long functions with multiple steps and large numbers
of smaller functions. I am deliberately choosing long functions as they
each fulfill a single contract. Furthermore, my aim in the near future
is to refactor the export logic away from a nightly big-bang approach
into hooks on individual models that continuously update BigQuery as the
records change. This will remove the need for this lib, but it unclear
when we might be able to schedule this work.