Change tables from a constant to an instance var
As a constant it causes load order errors with some rails commands, but
not others. For example, it works fine with `rails c`, but throws a
`NameError: Constant ApplicationRecord not found` when you try to run
`rails rspec` or `rails db:migrate`. Delaying the table load until
initialize has no consequences and solves this problem.