move sidebar to left of page and change sidebar contents
This change moves the sidebar to the left of the page and changes the
content of it to mainly be focused around key dates for vacancies e.g.
The application deadline. This also introduces a new days to apply prompt
to give jobseekers a quick overview of how many days they have left - done
with the introduction of a presenter method.

This also aims to reduce some of the tech debt around duplicated views
by pulling the sidebar into its own haml file that can be rendered, thus
reducing some of the maintenance overhead.