Refactor vacancies controller, refactor/pass tests
Refactored feature test to remove repetition.
Refactored hiring staff vacancies controller to use a private method for looking up a vacancy.
(It might be bad practice to let tests drive the design of a class, but
I found no other way to run a test, including trying a controller test.)
Stubbed vacancy lookup so that the controller uses tests' vacancy objects
to wrap the vacancy record, instead of creating its own vacancy object.
We need to use a `vacancy` object created in the test so that we can stub out the method
Vacancy#delete_documents, which otherwise will attempt HTTP connections.