Move filehandle creation until after records.none?
https://dfedigital.atlassian.net/browse/TEVA-553

This resolves the runtime problem of the BigQuery export on TEVA-553. As
I originally wrote it, the file gets written even if there are no
records. This results in a zero-length file being uploaded to the cloud
storage bucket and imported to BigQuery. BigQuery throws an exception
because a file with no content at all cannot contain a schema.

I have moved the filenhandle creation for each new file until after the
check for empty tables. If a table is empty, its file gets skipped.