Combine two param types into one for hiring_staff/vacancies_controller
Previously, #show, #edit and #delete used :id as the param key, while #review
and #summary used :job_id. This caused some duplication of methods in the
controller.
This commit puts #show, #edit and #delete in line with the param style for
all routes starting with /school/jobs/:job_id/.
It also updates all tests and other parts of the codebase in line with
this change.