Add styling from the prototype
This is caveated as 'not being production code' in a comment at the top
of the file as served by the prototyping toolkit. I'm checking it in
for now as it doesn't impress me as terrible and it is working. I'm
going to try the govuk_publishing_component scss it is based on next.
This appears to be production code. I didn't start with that as it is
obviously different and I didn't want to waste a lot of time trying to
make it work if something of "lower" quality worked.