Index Vacancy attributes for consumption by Algolia
To index Vacancy, run this in the console:
Vacancy.reindex

Vacancy.reindex! indexes Vacancy without deleting removed objects.

If on a free Algolia plan, you will be limited to a number of bytes
in the index, I think. In that case, try something like this:
Vacancy.where("slug < ?", "as").reindex
which indexes only those Vacancies where the slug attribute is
alphanumerically prior to 'as'.