Rename Algolia search attributes and use job_summary
We want to maximize the legibility of attributes in the Algolia dashboard
to be used by non-devs.
job_description has been replaced by job_summary in the schema
and we are truncating it to 256 characters to reduce the
size of records and to make it more likely
that searches will return useful results, rather than red
herrings hidden within long job_description s.