Change conversion to use correct time zone
This is to avoid the potential case where a Date is converted into
unix time with one time zone, and the unix time is converted back
to a date in a different time zone resulting in a different date.

Using midday means we don't have to worry about clocks going back.

Vacancy.updated_at uses a TimeWithZone object so it should be
handled separately.