Enable Algolia indexing on production
This re-enables the integration of the external service with our
app, now that we have moved to an environment which reliably
contains the correct environment variables.

We may not need to index on development environments. If we later
decide to, we can use:

algoliasearch per_environment: true, disable_indexing: Rails.env.test? do