Update GovUK PaaS docs
This adds a mention of how to avoid downtime when changing environment variables on the production environment, since I have done this twice in the past few days on production, and we need to document how to do this rather than relying on interpolating between Google results and our non-prod docs.

Since I've done this twice recently, it means we know that the instructions can work on production without breaking things, at least sometimes. But I don't really know what I'm doing, so would appreciate a careful review from Colin/Cesi to make sure this isn't missing something specific to the production environment.