Change haml stanzas to markdown block, adapt tests
The haml stanzas are not necessary as we do not need to have a
translation of this document. Furthermore, by breaking the document down
into multiple sections distributed over two files and built at runtime,
they make it very difficult to detect where sectional changes should be
made. This is especially true given that the typical source-of-truth for
our terms and conditions is normally a monolithic word doc.

That said, they were being used to drive two conditions on a single test, but
these can be done just as easily with simple text matchers.