Remove Algolia fall-back environment variables
The fall-back variables were introduced because when we used AWS as our
hosting solution, some instances did not load all the environment variables.

The problem seems to have gone away since GovUK PaaS migration.

We need all instances to have access to the real environment variables,
so that they can communicate with Algolia.