Remove failing js test
I have manually confirmed this is working on the running site (you need
to set the GOOGLE_TAG_MANAGER_CONTAINER_ID to ensure that failure
doesn't block this). I *suspect* the problem is one of timing, but
several different things might be happening.

This test should probably be re-implemented in the Jest test suite. I
will put in a ticket for this shortly. For now, I'm removing it as it is
not reflecting reality.