Enable recaptcha scoring, remove Auditor::Audit
As I noted in an earlier commit, all we are doing at this point is
collecting a score. Once we have some data, we can decide how we are
going to deal with failing scores.

I checked with Performance and we are not using the audit log at all as
it relates to feedback so I removed it.