Remove flakey test
This test fails depending on the run order. See 31245 reliably fails.

I *believe* this has something to with the timecop scope and ordering,
but haven't been able to get it to stop after multiple minor refactors.
However, the auditor is on the list for potential removal and we aren't
using the data anywhere at all according to Performance. Rather than
spend a lot of time yak shaving an over-broad, off-topic, low value
test, I'm pulling it instead.