Custom Cloudfront caching for 404 end 500 errors
By default it is 5 min. This has created a small downtime on 9/06.
We reduce it to avoid waiting for longer than expected.
We chose:
- 10s for 404 as 404 errors are served quickly by the app and we want to
be able to fix them quickly for all users
- 60s for 500 as there may be an internal problem with the app so we can
use cloudfront to reduce the load, in case the 500 is due to a load problem,
memory leak, database overloaded, etc. By increasing the caching time,
we may help to relieve the pressure on the app.