Remove algolia feature specs
Both of these run against the paid production instance and, at the time
of this ticket, consume 206 operations between them every time they are
run. I am removing them for now as this increases our operational costs
in an uncontrollable way.

TODO: add similar checks back in as *unit tests* that ensure the Algolia
rails client receives the expected calls without touching the remote
application.

NOTE: I decided against running them against a free app under our
Algolia account because, while getting to grips with the workflow, I was
able to completely consume the free-tier quota on my test app in just a
couple of hours. I did consider this. It is impractical as it displaces
the problem from the paid app to an unspecified number of free apps (as
we will have to continually add more as they run out of quota).