Use `#listed?` to filter indexed vacancies
This method, with a bit of tweaking to account for records missing the
required attributes, exactly duplicates the filter currently being
applied in the Algolia search builder class.

TODO: Set up a sidekiq job that either periodically calls `#reindex!`
*or* a sidekiq job that queues single-record releases at the correct
time. The latter approach will cut down on the number of unnecessary
Algolia operations but may be too complicated using the current
configuration. To be determined.