Disallow access to prod index from non-prod envs
All things being equal, allowing direct access to the production Algolia
index from non-production environments is high-risk at best. This change
ensure that the app will error if this is attempted.

It will require a change in working practices, but this should not be
too odious after a bit of practice.

TODO: Add the README section on how to reuse or create free tier indexes
for development and testing.