Remove test index naming logic
As we are moving to different apps to manage these types of
environments, and we can create those environments as needed, this logic
is no longer required.

TODO: Investiate operations overhead of managing these indexes here,
versus doing them solely in the dashboard. If they are switched to the
dashboard, then it might be a good idea to raise an exception here if
they get deleted. Alternately, I need to check the docs to check the
overhead of calling `add_replica` every time `reindex!` is called.