Drop login keys table (if exists) before creating
This is because the pre-prod environment staging was failing to restart
when I pushed this branch to it. It was failing with the error message
below. I have tested this on staging. I have also made the migrations
reversible in case a rollback is needed.

PG::DuplicateTable: ERROR:  relation "emergency_login_keys" already exists