Add function to target unindexed live records
This only targets records that are `live` but have not been indexed at
least once. The use case is records where the publication time has
passed, the expiry time has not yet arrived, the status is published,
but no further event has occurred on the record to cause it to be saved
and an index call thereby triggered.

A straight `.reindex!` under the circumstances is wasteful - it uses as
many algolia operations as there are records to be indexed even if no
records are updated. This ensures that records that have already been
indexed will not be mixed in with records that need to be indexed the
first time.