Remove algolia int test & note on changed keys
I may revisit this later depending on how the indexing work shapes up.
For now the plan is to check this using unit tests as this is unlikely
to alert us to changes in the Algolia API *before* we start getting
reports that the users are experiencing failures.

Also, I have changed the GitHub Actions config to use the Algolia free
tier app `TW5B7ETSIE`.