Teva 884 revised (#1660)
* Remove algolia feature specs

Both of these run against the paid production instance and, at the time
of this ticket, consume 206 operations between them every time they are
run. I am removing them for now as this increases our operational costs
in an uncontrollable way.

TODO: add similar checks back in as *unit tests* that ensure the Algolia
rails client receives the expected calls without touching the remote
application.

NOTE: I decided against running them against a free app under our
Algolia account because, while getting to grips with the workflow, I was
able to completely consume the free-tier quota on my test app in just a
couple of hours. I did consider this. It is impractical as it displaces
the problem from the paid app to an unspecified number of free apps (as
we will have to continually add more as they run out of quota).

* Use `#listed?` to filter indexed vacancies

This method, with a bit of tweaking to account for records missing the
required attributes, exactly duplicates the filter currently being
applied in the Algolia search builder class.

TODO: Set up a sidekiq job that either periodically calls `#reindex!`
*or* a sidekiq job that queues single-record releases at the correct
time. The latter approach will cut down on the number of unnecessary
Algolia operations but may be too complicated using the current
configuration. To be determined.

* Disallow access to prod index from non-prod envs

All things being equal, allowing direct access to the production Algolia
index from non-production environments is high-risk at best. This change
ensure that the app will error if this is attempted.

It will require a change in working practices, but this should not be
too odious after a bit of practice.

TODO: Add the README section on how to reuse or create free tier indexes
for development and testing.

* Remove test index naming logic

As we are moving to different apps to manage these types of
environments, and we can create those environments as needed, this logic
is no longer required.

TODO: Investiate operations overhead of managing these indexes here,
versus doing them solely in the dashboard. If they are switched to the
dashboard, then it might be a good idea to raise an exception here if
they get deleted. Alternately, I need to check the docs to check the
overhead of calling `add_replica` every time `reindex!` is called.

* Override default Algolia reindex calls

This ensures correct targeting without having to rely on people
remembering to include `.live` to scope their calls.  It probably
obviates the need for the `if: :listed?` conditionial in *ordinary*
operations.

It also adds back in `auto_index` and `auto_remove`.

TODO:
* Add notes to README on how operations work re. creating replica
  indexes from the model (i.e. no extra overhead and ensure they're
  always there after an index).
* Still need to work out how to manage records that hit their
  publication date at some point in the future. This is because adding
  records to the index, even with `auto_index` is event-driven.  A
  really simple way would probably be just to run a cron job that checks
  their dates and touches them if they've arrived at their
  to-be-published date. That would make the index updates atomic (i.e.
  would not have to check every record as you would with a `.reindex!`
  call). Will expereiment tomorrow and put it in if it works as
  expected.

* Remove test-specific index naming conventions

We are no longer specially adapting algolia indexes for the test
environements, so these are no longer needed.

* Add function to target unindexed live records

This only targets records that are `live` but have not been indexed at
least once. The use case is records where the publication time has
passed, the expiry time has not yet arrived, the status is published,
but no further event has occurred on the record to cause it to be saved
and an index call thereby triggered.

A straight `.reindex!` under the circumstances is wasteful - it uses as
many algolia operations as there are records to be indexed even if no
records are updated. This ensures that records that have already been
indexed will not be mixed in with records that need to be indexed the
first time.

* Do not exclude algolia tagged specs

As they have all been removed.

* Make `#listed?` use `expiry_time` & tolerate nulls

This was using `expires_on`, which is a Date object while `expiry_time`
is a DateTime object. See comment for interesting bug around nullable
DateTime attributes and negation.

TODO:
 * Refactor scopes to use `expiry_time` instead of combinations of
   `expires_at` and `expires_on`.

* Add a method to call unscoped `Vacancy#reindex!`

See comment for use case an removal instructions.

* Alphabetize

* Add Algolia info and remove some out-of-date info

* Add sidekiq cron to make published vacancies live

I decided to start with a five minute interval although it *could* be
shorter. Review it in a few weeks to see if there are any problems.

* Add method to delete yesterday's expired records

* Set up cron jobs to remove expired records

* Add docs on env variables and indexing calls

* Restore Algolia key setup to test environment.

* Remove algolia int test & note on changed keys

I may revisit this later depending on how the indexing work shapes up.
For now the plan is to check this using unit tests as this is unlikely
to alert us to changes in the Algolia API *before* we start getting
reports that the users are experiencing failures.

Also, I have changed the GitHub Actions config to use the Algolia free
tier app `TW5B7ETSIE`.

* Re-add missing job spec

I only meant to remove the bit that specifically called the algolia
index, but killed the entire file instead. This is it without the
algolia call but with the logic that ensures old-style alerts still
work.

* Remove transitional tests

As per @cpjmcquillan's comment.

* Add note on free algolia apps