Restore CSRF protection for job alert form (#1708)
This introduces an attack vector that is riskier than running with
the current bug (TEVA-881). I am overriding this release and will
investigate the cause in more detail. A quick glance at Stack Overflow
suggests a few different ways this might be dealt with without disabling
CSRF protection on our busiest form.

Details of CSRF are available:x
[here](https://guides.rubyonrails.org/security.html#cross-site-request-forgery-csrf),
[here](https://owasp.org/www-community/attacks/csrf), and
[here](https://owasp.org/www-community/attacks/csrf)