Make tests more robust
These tests had been using an interpolated value (vacancy.job_location)
immediately after testing an interaction with a feature.

If the interaction didn't have the correct effect on vacancy.job_location
(for example if it set it to wrong_value) then these tests would not have
caught the error.

We should explicitly define what we want the result to be, in general, rather
than generating it dynamically. So I have defined the expected translation
labels.