Fix flaky tests
In short, apostrophes are encoded for html pages, making feature tests
appear to fail occasionally when a school name is for example O'Reilly
School. This results in a check whether O'Reilly School (the string)
is included in a page which renders O &#39; Reilly School.

As these all come from the factories, we can just nip them in the bud.

For more context, see this blog post:

https://medium.com/@randallreedjr/intermittent-test-failures-with-factorygirl-faker-and-apostrophes-b47a26288fc4