Include method for deleting expired records since year start
This is the method I used to clear expired records from the index which had
been missed.

Because I had to play around with it a little (e.g. had to use batches) it took about
20 minutes to get going.

It may be worth including this method in vacancy.rb to save time if we have to do this
again in the future.

It also might not be worth the maintenance cost / space considerations.