Randomly add assignees to Dependabot PRs
This is to avoid the situation where nobody reviews the PR because it's too boring, and then the most altruistic person has to review all Dependabot PRs, and the PRs sit there saying 'high security issue' for weeks.

The auto-assignment bot will trigger unless the PR title contains `wip` or `teva`.

Unfortunately, this bot will trigger on PRs which are made by us humans, but don't contain those keywords, like this one doesn't. We could agree either:

(1) to add a keyword to all human PR titles that don't include `teva`, e.g. `*` (and add this to the list to skip)

or (2) ignore the assignee in these cases.