Do not omit nil vacancy attributes when merging to session
If we omit 'nil' attributes, then no overwriting of previous values happens.

We want to overwrite previous session attributes when the user goes back a step
and reselects a different option.

For example, when selecting 'at one school' on Step 1, the readable_job_location
should be set to nil, otherwise the previous value will be used instead in
the page heading ('Create a job listing at More than one school ()', in this
case.)